代码走查

+++ src/main/java/ttjj/dto/RankBizDataDiff.java
@@ -118,7 +118,7 @@ public class RankBizDataDiff {
     /**
      * f21 流通市值
      */
-    private Long f21;
+    private BigDecimal f21;
     private Long f22;
     /**
      * f23 市净率
@@ -647,11 +647,11 @@ public class RankBizDataDiff {
         this.f20 = f20;
     }
 
-    public Long getF21() {
+    public BigDecimal getF21() {
         return f21;
     }
 
-    public void setF21(Long f21) {
+    public void setF21(BigDecimal f21) {
         this.f21 = f21;
     }
